Refactor Token auth to support Token refreshes.
Adds the ability to refresh a token if the token expires and it has a
refresh token. This allows us to use the FHIR Client's dynamic `auth_callback`
approach, which now checks to see if the token needs to refresh, refreshes it
if needed, before creating the HTTP Authorization header.

Resolves #9.